+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## XX.XX.XX - 2023-XX-XX
 
+* [Added] Added support for `paymentMethodTypes` in `CustomerAdapter`, which filters which payment methods are shown in the add screen.
+
 ## 20.38.0 - 2024-02-26
 
 ### PaymentSheet